Move P2P configuration into proper config part
